moved layout to container
